+++ CHANGELOG.md
@@ -3,5 +3,12 @@ All notable changes to the "lightcode" extension will be documented in this file
 
 Check [Keep a Changelog](http://keepachangelog.com/) for recommendations on how to structure this file.
 
-## [Unreleased]
-- Initial release
+## [0.0.2]
+### Added
+- Light Code server.
+- Hover provider.
+- Definition provider.
+
+
+## [0.0.1]
+- Initial release.